requirements.txt
anthropic
python-telegram-bot
python-dotenv
psutil
